+++ src/main/java/com/ssafy/ssafsound/infra/storage/dto/PostStoreImageResDto.java
@@ -3,11 +3,9 @@ package com.ssafy.ssafsound.infra.storage.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.util.List;
-
 @Getter
 @AllArgsConstructor
 public class PostStoreImageResDto {
 
-    private List<ImagePathDto> imagePathDtos;
+    private ImageUploadDto imageUploadDto;
 }
